requirements.txt
speechrecognition
pyaudio
gTTS
RPi.GPIO
pymongo
dnspython
python-decouple==3.3
